requirements.txt
# Core Web Framework
Flask==2.3.3
Werkzeug==2.3.8
Jinja2==3.1.3
MarkupSafe==2.1.4
itsdangerous==2.1.2
click==8.1.7

# Flask Extensions
flask-cors
flask-limiter

# ML / Deep Learning
torch==2.5.1
torchvision==0.20.1
numpy==1.26.4
scipy==1.13.1
Pillow==10.2.0

# Computer Vision
opencv-python==4.9.0.80

# Utilities
tqdm==4.66.1
python-dotenv
typing-extensions==4.9.0
filelock==3.13.1

# Testing
pytest